Track transaction outputs
Necessary to handle the edge case if a transactions pays to two separate
invoices, or twice to the same invoice address.

Don't track transactions in database, only the tx outputs, as it's a
simpler design and we can rebuild the transaction if we ever need it
(but I don't see a use-case for it now).