Basic server setup, routing to an api or web context
At first this repo was supposed to be just an Elixir library,
and the server part would use the library, but I don't really see a use-case
for the library other than as the backend for the library.

It also needlessly complicates the development and when we want to add a
connection between invoices and stores (in the future) having a separate
library for the core is just silly.